refactor: 리뷰 반영
createdAt not null
Test 이름 변경
