+++ src/test/java/qna/repos/UserRepositoryTest.java
@@ -1,10 +1,10 @@
-package qna.domain;
+package qna.repos;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import qna.repos.UserRepository;
+import qna.domain.User;
 
 import java.util.Optional;
 
@@ -12,14 +12,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 @DataJpaTest
-public class UserTest {
+public class UserRepositoryTest {
     public static final User JAVAJIGI = new User(1L, "javajigi", "password", "name", "javajigi@slipp.net");
     public static final User SANJIGI = new User(2L, "sanjigi", "password", "name", "sanjigi@slipp.net");
 
     @Autowired
     private UserRepository repository;
 
-    @DisplayName("Answer 저장 테스트")
+    @DisplayName("User 저장 테스트")
     @Test
     void save() {
         User expected = JAVAJIGI;
